requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
numpy==1.26.4
tensorflow==2.15.0
soundfile==0.12.1
dj-database-url==2.1.0